Nikita/setting navigation

+++ app/src/main/java/ru/itis/looktomoney/domain/Change.kt
@@ -1,13 +1,13 @@
 package ru.itis.looktomoney.domain
 
 class Change {
-    var numb : Int = 0
+    var numb : Double = 0.0
     var category : Category? = null
     var wallet : Wallet? = null
 
     constructor()
 
-    constructor(int: Int, category: Category, wallet: Wallet){
+    constructor(int: Double, category: Category, wallet: Wallet){
         numb = int
         this.category = category
         this.wallet = wallet
